Interface

Module

OpenFIGI — Module

Functionality to access the OpenFIGI API

See the python example.

Interface

OpenFIGI.set_apikey — Function

set_apikey(apikey)

Sets the (global) apikey to apikey

OpenFIGI.cache_enums — Function

cache_enums(; refresh=false)

Caches enumeration values in a global dictionary for quick access. If refresh is true, forces the enumeration values to be queried from the API instead of loaded from the scratch space.

Rate Limits

OpenFIGI.mapping_channel — Function

mapping_channel()

Creates a new channel for mapping tasks

The resulting Channel{Task} instance is sized to hold one interval of request tasks and is bound to a new Task spawned to rate-limit the channel.

mapping_channel(f, wait_reset=false)

Creates a channel for mapping tasks, runs f on that channel, and finally closes the channel.

nvda = search_channel() do ch
    return mapping(ch, Ticker("NVDA"), ExchCode("US"), SecurityType2("Common Stock"))
end

See mapping_channel()

OpenFIGI.search_channel — Function

search_channel()

Creates a channel for search or filter tasks

The resulting Channel{Task} instance is sized to hold one interval of search or filter tasks and is bound to a new Task spawned to rate-limit the channel.

search_channel(f, wait_reset=false)

Creates a channel for search or filter tasks, runs f on that channel, and finally closes the channel.

See search_channel()