Search & Filter

Rate-Limited Methods

OpenFIGI.Search.instrument_search — Method

instrument_search(tasks, query, properties...)

Submits query to the search endpoint as scheduled by tasks

Arguments

tasks::Channel{Task}: a rate-limited channel for scheduling API requests (see OpenFIGI.search_channel())

query::String: search keywords

properties::AbstractProperty...: optional properties for limiting results

Internally, this method uses instrument_search(tasks, results, query, properties...) for the batching logic.

OpenFIGI.Search.instrument_search — Method

instrument_search(tasks, results, query, properties...)

Spawn a instrument_search task that processes query in batches and populates results

Arguments

tasks::Channel{Task}: a rate-limited channel for scheduling API requests (see OpenFIGI.search_channel())

results::Channel{Instrument}: the search results in order of return

query::String: search keywords

properties::AbstractProperty...: optional properties for limiting results

The task returned from this method is bound to results.

OpenFIGI.Filter.instrument_filter — Method

instrument_filter(tasks, query, properties...)

Submits query to the filter endpoint as scheduled by tasks

Arguments

tasks::Channel{Task}: a rate-limited channel for scheduling API requests (see OpenFIGI.search_channel())

properties::AbstractProperty...: optional properties for limiting results

query::String: search keywords

Internally, this method uses instrument_filter(tasks, results, properties...; query) for the batching logic.

OpenFIGI.Filter.instrument_filter — Method

instrument_filter(tasks, results, properties...; [query])

Spawn a instrument_search task that processes query in batches and populates results

Arguments

tasks::Channel{Task}: a rate-limited channel for scheduling API requests (see OpenFIGI.search_channel())

results::Channel{Instrument}: the search results in order of return

properties::AbstractProperty...: optional properties for limiting results

query::String: optional filter keywords

The task returned from this method is bound to results.

Direct Methods

OpenFIGI.Search.instrument_search — Method

instrument_search(query, properties...)

Sends a single search query to the OpenFIGI API

Caveats

This method is primarily for testing.

This method does not support paginated responses, so no more than 100 results are returned.

This method does not attempt to manage requests to comply with rate limits.

Use instrument_search(tasks, query, properties...) instead.

OpenFIGI.Filter.instrument_filter — Method

instrument_filter(query, properties...)

Sends a single filter query to the OpenFIGI API

Caveats

This method is primarily for testing.

This method does not support paginated responses, so no more than 100 results are returned.

This method does not attempt to manage requests to comply with rate limits.

Use instrument_filter(tasks, query, properties...) instead.